Geisha System Notifications set to Fixed Position An example of notification bar set in a fixed position for reference
- updated geisha.css

At the moment I only updated the notificiation bar in the Geisha layout so it should reflect an example of how the notification bar should look and act for other templates. The CSS is a little funky but the method of setting Bottom and Right and leaving out Top makes it so the element when resized only increase in height upward instead of gaining size downward when there's a lot of text.

Currently the change has only been made to the Geisha layout because of a display issue in 2 known areas. Mod Page and Settings Page. When there is an system message displayed there, the system message will not be displayed in its usual spot it'll be echo'd out just a few blocks underneath causing a display issue. The current fix is unknown but i'm guessing right now after 4 minutes of browsing around, the fix might have to do with some refactoring but tbh atm idk.